ref(ct:api): move responses viewset to api from views
Changes:
- changes api url to `api/courses/<course_id>/responses/`
